attached_assets/requirements_1754544163487.txt
fastapi==0.68.1
uvicorn==0.15.0
websockets==9.1
python-binance==1.0.15
aiohttp==3.8.1
redis==4.5.1
python-dotenv==0.19.2
pydantic==1.10.7